Ready driver for motor enable
The driver now accepts initialization messages containing "motoractive"
names.

However, the state of the motoractives are not forwarded and the enable
signals are not sent to the robot.

Further, even though the positions of the arm is not setable, the driver
still sends setpoints.

A diagnostics updater has been included in the ros driver class to take
care of relaying the info about the motor drives.